Don't stop nginx if it isn't running
In some cases nginx should be started and stopped by the system it self.
When the user already has shut it down there is no point trying to kill it.
Since sudo behaves in an annoying way on some configurations running it when
there is no need to should be avoided.

+++ scripts/stop-freeciv-web.sh
@@ -7,7 +7,9 @@ export FREECIV_WEB_DIR="${SCRIPT_DIR}/.."
 echo "Shutting down Freeciv-web: nginx, resin, publite2, freeciv-proxy."
 
 # 1. nginx
-sudo killall nginx
+if [ "$(pidof nginx)" ] ; then
+  sudo killall nginx
+fi
 
 # 2. Resin
 ${FREECIV_WEB_DIR}/resin/bin/resin.sh stop
